# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Doctors Scheduler",
    "version": "1.0.0"
  },
  "schemes": [
    "http",
    "https"
  ],
  "basePath": "/api/v1",
  "paths": {
    "/doctors": {
      "get": {
        "summary": "Get all doctors",
        "description": "Retrieves the list of doctors",
        "responses": {
          "200": {
            "description": "The list of doctors",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Doctor"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a doctor",
        "parameters": [
          {
            "name": "doctor",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NewDoctor"
            }
          }
        ],
        "description": "Creates a single doctor",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The newly created doctor",
            "schema": {
              "$ref": "#/definitions/Doctor"
            }
          }
        }
      }
    },
    "/doctors/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "summary": "Get single doctor",
        "description": "Get doctor by id",
        "responses": {
          "200": {
            "description": "Single doctor",
            "schema": {
              "$ref": "#/definitions/Doctor"
            }
          }
        }
      }
    },
    "/patients": {
      "get": {
        "summary": "Get all patients",
        "description": "Retrieves the list of patients",
        "responses": {
          "200": {
            "description": "The list of patients",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Patient"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a patient",
        "parameters": [
          {
            "name": "patient",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NewPatient"
            }
          }
        ],
        "description": "Creates a single patient",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The newly created patient",
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        }
      }
    },
    "/patients/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "summary": "Get single patient",
        "description": "Get patient by id",
        "responses": {
          "200": {
            "description": "Single patient",
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        }
      }
    },
    "/visits": {
      "get": {
        "summary": "Get visits filtered by doctor id, patient id, both, or none",
        "description": "Retrieves list of filtered visits",
        "parameters": [
          {
            "name": "doctorid",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "patientid",
            "in": "query",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "The list of filtered visits",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Visit"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Doctor": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "NewDoctor": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Patient": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "NewPatient": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Visit": {
      "type": "object",
      "properties": {
        "doctorId": {
          "type": "integer"
        },
        "patientId": {
          "type": "integer"
        },
        "date": {
          "type": "string"
        }
      }
    }
  }
}